Begin shift to new rpc structure
