+++ src/index.ts
@@ -1,5 +1,5 @@
-import { createChainRpc } from './chainrpc';
 import { createLnRpc } from './lnrpc';
+import { createChainRpc } from './rpc/chain-rpc';
 
 export { createChainRpc };
 export default createLnRpc;
